app/requirements.txt
# FastAPI and web framework
fastapi==0.115.13
uvicorn[standard]==0.34.3
pydantic==2.11.7
pydantic-settings==2.9.1

# Database and ORM
sqlalchemy==2.0.41
alembic==1.16.2
psycopg2-binary==2.9.10
asyncpg==0.30.0

# Vector database and embeddings
sentence-transformers==4.1.0
faiss-cpu==1.11.0
numpy==2.3.0

# Document processing
pdfplumber==0.11.7
python-docx==1.2.0
striprtf==0.0.26
chardet==5.2.0
Pillow==11.2.1

# Background tasks
celery==5.5.3
redis==6.2.0

# Utilities
python-multipart==0.0.20
aiofiles==24.1.0
httpx==0.28.1
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
tiktoken==0.8.0

# Development and testing
pytest==8.4.1
pytest-asyncio==0.24.0
pytest-mock==3.14.0
pytest-cov==5.0.0
pytest-xdist==3.6.0
factory-boy==3.3.1
faker==30.8.2
black==23.12.1
flake8==7.3.0
aiosqlite==0.20.0

# Monitoring and logging
structlog==25.4.0

vertexai==1.71.1